+++ src/main/java/dgraph/Test.java
@@ -4,6 +4,8 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Map;
 
+import dgraph.node.Company;
+import dgraph.node.Industry;
 import dgraph.node.Label;
 import dgraph.node.NodeUtil;
 import dgraph.node.School;
@@ -19,7 +21,6 @@ public class Test {
   Demo demo = new Demo(dClient);
 
   public void test_one() {
-    demo.init();
     School school = new School();
     String name = "清华大学";
     String type = "学校";
@@ -47,10 +48,47 @@ public class Test {
     dClient.entityAddAttrTest("分析化学", "has_label", "0x118d");
   }
 
+  private void test_tree() {
+    Label label = new Label();
+    label.setLabel_name("公司类型");
+    // "公司类型": "0x118b"
+    label.setUid("0x15");
+    Industry industry = new Industry();
+    String industryType = "行业";
+    String industryName = "互联网";
+    String industryCode = "8001";
+    industry.setType(industryType);
+    industry.setUid("0x13");
+    industry.setUnique_id(industryName);
+    industry.setCode(Integer.parseInt(industryCode));
+    industry.setName(industryName);
+    Company company = new Company();
+    String name = "腾讯有限公司";
+    String location = "深圳";
+    String type = "公司";
+    company.setName(name);
+    company.setUnique_id(name);
+    company.setLocation(location);
+    company.setType(type);
+    company.setIndustry(Arrays.asList(industry));
+    label.setCompany(company);
+    Map<String, String> industryEntityUidMap = NodeUtil.insertEntity(dClient, label.getCompany().getIndustry());
+    FileUtils.saveFile("src/main/resources/test_industry_uid_map.txt", industryEntityUidMap);
+    NodeUtil.putEntityUid(label.getCompany().getIndustry(), industryEntityUidMap);
+    Map<String, String> companyEntityUidMap = NodeUtil.insertEntity(dClient, Arrays.asList(label.getCompany()));
+    NodeUtil.putEntityUid(Arrays.asList(label.getCompany()), companyEntityUidMap);
+    FileUtils.saveFile("src/main/resources/test_company_uid_map.txt", companyEntityUidMap);
+    NodeUtil.putEntityUid(Arrays.asList(label.getCompany()), companyEntityUidMap);
+    Map<String, String> uid = NodeUtil.insertEntity(dClient, Arrays.asList(label));
+    FileUtils.saveFile("src/main/resources/test_label_uid_map.txt", uid);
+  }
+
   public static void main(String[] arg) {
     Test test = new Test();
+    // test.demo.init();
+    // test.test_one();
     // test.test_two();
-    test.test_one();
+    test.test_tree();
 
   }