+++ src/main/java/dgraph/node/Label.java
@@ -15,6 +15,9 @@ public class Label extends EntityNode {
 
   List<School> school;
 
+  Company company;
+
+  Industry industry;
 
   public List<School> getSchool() {
     return school;
@@ -22,5 +25,21 @@ public class Label extends EntityNode {
 
   public void setSchool(List<School> school) {
     this.school = school;
+  }
+
+  public Company getCompany() {
+    return company;
+  }
+
+  public void setCompany(Company company) {
+    this.company = company;
+  }
+
+  public Industry getIndustry() {
+    return industry;
+  }
+
+  public void setIndustry(Industry industry) {
+    this.industry = industry;
   }
 }